Moves tracking tests
