+++ packages/tracking/__tests__/android/jest.config.js
