+++ packages/tracking/__tests__/track-events.test.js
@@ -2,7 +2,7 @@ import React from "react";
 import { Text } from "react-native";
 import renderer from "react-test-renderer";
 import PropTypes from "prop-types";
-import { withTrackEvents } from "../tracking";
+import { withTrackEvents } from "../src/tracking";
 import withTrackingContext from "./test-tracking-context";
 import sharedTrackingTests from "./shared-tracking-tests";
